Fixed TruckingDoor audio playing when door is already opened/closed
Fixed bug where, if the door was open and the player tried to open the door, the audio would play regardless. The same thing occurred for closing the door. A simple check was put in place to confirm if the door was already in the state it was trying to be changed to, it will not play the sound.