Changed Main Background, Added Sounds, and Started Sorting
Changed the Main Desktop Background to allow for coffee steam animations.

Started adding some sorting stuff in, copied trucking camera/objects and changed names/background.

Of note, sounds are now able to be used in the game! There is a new GameObject and Script called DialogueAudioPlayer. I have altered the Dialogue class to contain a list for sounds, which are ints.

It works very similar to the way that expressions and names do. The DialogueManager class will make a Queue of all the sounds, and will Dequeue one every sentence. If there is not one to Dequeue, it will give the code for "No Sound"

If you look in the DialogueAudioPlayer script, in the Start() code it will load all the sounds that it needs. If we add new sounds, the current ones can be used as a template for how to define them. I found the PlayOneShot() function for AudioSources, and they are wonderful.