Fixed Bexos and Unknown Dialogue Portraits and Other Bugs
Added in new Bexos and Unkown dialogue box portraits, and ensured switching between them worked.

Modified DialogueManager to enqueue a "None sound" if there is no value from the line to enter, as this ensures that sounds will play along with the correct dialogue line.

Fixed minor error in Entry Conversation where I.V.A.A.L.'s nameplate was written as '5'

Switched the timer for the conversations back to 45, to make sure that things can't be sped through.