Testing Out Splitting of Queues, and Implemented all Conversations
Split up the Queues for Chapter 2 into multiple queues to allow for easy staging of tasks.
Imported all Conversation Objects into Chapter 2, so you should be able to make it the whole way through (no exit for Chapter 2 yet, do that in TriggerManager)
Seems like there's still an issue with Tasks being queued right out of the gate for Chapter2, have to look into that.